improve and complete feature testing
