+++ src/Serilog.Ui.Web/src/app/hooks/useSearchForm.ts
@@ -1,6 +1,4 @@
 import { isArrayGuard } from 'app/util/guards';
-import { parseSearchParams } from 'app/util/queryParams';
-import { useEffect } from 'react';
 import { useForm, useFormContext } from 'react-hook-form';
 import { useSearchParams } from 'react-router';
 import {
@@ -45,7 +43,7 @@ export const useSearchForm = () => {
     defaultValues: searchFormInitialValues,
   });
   const useSearchContext = useFormContext<SearchForm>();
-  const [searchParams, setSearchParams] = useSearchParams();
+  const [, setSearchParams] = useSearchParams();
 
   const { data } = useQueryTableKeys();
   const tableKeysDefaultValue = isArrayGuard(data) ? data.at(0)! : '';
@@ -56,43 +54,14 @@ export const useSearchForm = () => {
       tableKeysDefaultValue,
     );
 
-    const tableValue = !blankTable
-      ? new URLSearchParams({ table: tableKeysDefaultValue })
-      : '';
+    const tableValue =
+      !blankTable && tableKeysDefaultValue
+        ? new URLSearchParams({ table: tableKeysDefaultValue })
+        : '';
     setSearchParams(tableValue);
 
     return runRefetch;
   };
-
-  useEffect(() => {
-    const { urlParams, cleanedFromInvalidQueryString } = parseSearchParams(searchParams);
-    const urlParamKeys = Object.keys(urlParams);
-    const currentAppValues = methods.getValues();
-
-    // we set values for anything coming from the URL
-    urlParamKeys.forEach((e) => {
-      const paramExists = e in currentAppValues;
-      if (!paramExists) return;
-
-      if (urlParams[e] === currentAppValues[e]) return;
-      methods.setValue(e as keyof SearchForm, urlParams[e]);
-    });
-
-    // if the URL doesn't have the key, we need to remove it from the application search
-    const resetMissingValues = Object.keys(currentAppValues).filter(
-      (value) => !urlParamKeys.includes(value),
-    );
-    resetMissingValues.forEach((element) => {
-      methods.setValue(element as keyof SearchForm, searchFormInitialValues[element]);
-    });
-
-    // if there were invalid parameters, we need to set a cleaned query string
-    if (!cleanedFromInvalidQueryString) return;
-    setSearchParams(cleanedFromInvalidQueryString);
-
-    // no need to register the useEffect for methods/setSearchParams
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [searchParams]);
 
   return { methods, ...useSearchContext, reset: resetForm };
 };
